@@ -21,15 +21,9 @@ final class GetRectorKernel extends Kernel
 {
     use MicroKernelTrait;
 
-    /**
-     * @var FlexLoader
-     */
-    private $flexLoader;
+    private FlexLoader $flexLoader;
 
-    /**
-     * @var Discovery
-     */
-    private $discovery;
+    private Discovery $discovery;
 
     public function __construct(string $environment, bool $debug)
     {
