@@ -9,10 +9,7 @@ use Rector\Website\Entity\ResearchAnswer;
 
 final class ResearchAnswerRepository
 {
-    /**
-     * @var EntityManagerInterface
-     */
-    private $entityManager;
+    private EntityManagerInterface $entityManager;
 
     public function __construct(EntityManagerInterface $entityManager)
     {
